[android] Move bookmark subscription ui to purchase package
